moving  CoachExamsPage
