+++ kolibri/plugins/coach/assets/src/views/exams/ExamReportPage/ManageExamModals.vue
@@ -62,7 +62,7 @@
   import { mapState, mapActions, mapGetters } from 'vuex';
   import xorWith from 'lodash/xorWith';
   import AssignmentChangeStatusModal from '../../assignments/AssignmentChangeStatusModal';
-  import ExamPreview from '../CoachExamsPage/ExamPreview';
+  import ExamPreview from '../../plan/CoachExamsPage/ExamPreview';
   import AssignmentDetailsModal from '../../assignments/AssignmentDetailsModal';
   import AssignmentCopyModal from '../../assignments/AssignmentCopyModal';
   import AssignmentDeleteModal from '../../assignments/AssignmentDeleteModal';
