+++ kolibri/plugins/coach/assets/src/views/plan/CoachExamsPage/index.vue
@@ -147,8 +147,8 @@
   import { PageNames } from '../../../constants';
   import imports from '../../new/imports';
   import PlanHeader from '../../new/PlanHeader';
-  import ManageExamModals from '../ExamReportPage/ManageExamModals';
-  import ExamReportPage from '../ExamReportPage';
+  import ManageExamModals from '../../exams/ExamReportPage/ManageExamModals';
+  import ExamReportPage from '../../exams/ExamReportPage';
   import AssignmentDetailsModal from '../../assignments/AssignmentDetailsModal';
   import AssignmentCopyModal from '../../assignments/AssignmentCopyModal';
   import AssignmentDeleteModal from '../../assignments/AssignmentDeleteModal';
